[new release] travesty (0.6.0) (#14355)
* [new release] travesty (0.6.0)

CHANGES:

Major release with various breaking changes and new features, the main one
being bi-traversables.

## Breaking changes

- Most module signatures have moved from `XYZ` to `XYZ_types`.  For example,
  `Bi_mappable.S2` is now `Bi_mappable.S2`.  This is to eliminate
  the 'intf pattern' previously used in Travesty.
- Removed `Travesty_core_kernel_exts`.  Use `Travesty_base_exts`
  instead.
- Made `Bi_mappable.Make2`'s return module type use sharing constraints
  instead of destructive substitutions.  This may cause shadowing where there
  previously wasn't any.

## Bi_traversable

This release adds a `Bi_traversable` module---effectively being to
`Traversable` as `Bi_mappable` is to `Mappable`.

`Alist`, `Tuple2`, and `Or_error` now expose `Bi_traversable.S2`, which
subsumes their previous interface.

## Biffs

This release also adds the ability to compose `Mappable`s on the inside of
`Bi_mappable`s (and `Traversables` on the inside of `Bi_traversables`): what
is often referred to as a 'biff' in the quirky world of Haskell.  This
complements the ability to compose them on the outside (a 'tannen').

## Minor improvements

- Removed spurious dependencies that caused travesty to be unavailable on
  OCaml 4.08.
- `Bi_mappable` arity-1 chaining functors now carry previously-missing
  sharing constraints equating their fixed types with those of their
  bi-mappable ancestor.
- Added experimental F#-style `>>` operator in
  `Travesty_base_exts.Fn.Compose_syntax`.
- Added bi-traversable instance for `Result`, from which `Or_error` now
 inherits.

* Update opam